Engram-AiMemory/01_devroot/RemoteAgent/agent/pyproject.toml
[project]
name = "embedding-agent"
version = "0.1.0"
description = "Local embedding agent that streams to remote Weaviate"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mlx>=0.21.0",
    "mlx-embedding-models>=0.0.11",
    "watchdog>=4.0.0",
    "weaviate-client>=4.9.0",
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "rich>=13.9.0",
    "typer>=0.15.0",
    "xxhash>=3.5.0",
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "tiktoken>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
lmstudio = [
    "lmstudio>=1.0.0",
]

[project.scripts]
embedding-agent = "embedding_agent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
